+++ src/test/kotlin/blackjack/domain/card/CardTest.kt
@@ -1,7 +1,12 @@
-package blackjack.domain
+package blackjack.domain.card
 
-import blackjack.domain.Denomination.*
-import blackjack.domain.Suit.SPADE
+import blackjack.domain.card.Denomination.ACE
+import blackjack.domain.card.Denomination.JACK
+import blackjack.domain.card.Denomination.KING
+import blackjack.domain.card.Denomination.QUEEN
+import blackjack.domain.card.Denomination.THREE
+import blackjack.domain.card.Denomination.TWO
+import blackjack.domain.card.Suit.SPADE
 import io.kotest.core.spec.style.FunSpec
 
 class CardTest : FunSpec({
